Add hw7

+++ src/CTDL/Homework_7/LineSegmentIntersection/Point.java
@@ -3,13 +3,13 @@
 package CTDL.Homework_7.LineSegmentIntersection;
 
 public class Point {
-    public double x = 0, y = 0;
+    public int x = 0, y = 0;
 
     public Point() {
 
     }
 
-    public Point(double x, double y) {
+    public Point(int x, int y) {
         this.x = x;
         this.y = y;
     }
